+++ packages/zosfiles/src/api/methods/get/Get.ts
@@ -12,10 +12,8 @@
 import { posix } from "path";
 import { AbstractSession, Logger, ImperativeExpect } from "@zowe/imperative";
 import { ZosFilesMessages } from "../../constants/ZosFiles.messages";
-import { ZosmfHeaders } from "../../../../../rest/src/ZosmfHeaders";
-import { IHeaderContent } from "../../../../../rest/src/doc/IHeaderContent";
 import { ZosFilesConstants } from "../../constants/ZosFiles.constants";
-import { ZosmfRestClient } from "../../../../../rest";
+import { ZosmfRestClient, ZosmfHeaders, IHeaderContent } from "../../../../../rest";
 import { IGetOptions } from "./doc/IGetOptions";
 import { ZosFilesUtils } from "../../utils/ZosFilesUtils";