+++ packages/zosmf/__tests__/cli/check/status/Status.handler.unit.test.ts
@@ -15,7 +15,7 @@ import CmdHandler from "../../../../src/cli/check/status/Status.handler";
 import * as cmdDef from "../../../../src/cli/check/status/Status.definition";
 import { getMockedResponse, UNIT_TEST_ZOSMF_PROF_OPTS, UNIT_TEST_PROFILES_ZOSMF } from "../../../../../../__tests__/__src__/mocks/ZosmfProfileMock";
 
-jest.mock("../../../../src/api/methods/CheckStatus");
+jest.mock("../../../../src/api/CheckStatus");
 
 const goodCmdParms: IHandlerParameters = {
     arguments: {
@@ -23,6 +23,7 @@ const goodCmdParms: IHandlerParameters = {
         _: ["zosmf", "check", "status"],
         ...UNIT_TEST_ZOSMF_PROF_OPTS
     },
+    positionals: ["zosmf", "check", "status"],
     response: getMockedResponse(),
     definition: cmdDef.StatusDefinition,
     fullDefinition: cmdDef.StatusDefinition,