+++ CHANGELOG.md
@@ -6,6 +6,12 @@ All notable changes to the Zowe CLI package will be documented in this file.
 
 - Added API to delete migrated data sets.
 
+## `6.16.0`
+
+- Upgrade Zowe commands to prompt for any of the following values if the option is missing: host, port, user, and password.
+- Add ability to log into and out of the APIML, getting and using a token
+- Add `--base-profile` option to all commands that use profiles, allowing them to make use of base profiles containing shared values.
+
 ## `6.15.0`
 
 - Add `encoding` option to `zosmf` profile type.